Working with a kdevinecop object

A vine copula density estimate (stored in a kdevinecop object) can be evaluated on arbitrary points with dkevinecop. Furthermore, you can simulate from the estimated density with rkdevinecop.

dkdevinecop(u, obj, stable = FALSE)

rkdevinecop(n, obj, U = NULL, quasi = FALSE)

Arguments

u

\(m x 2\) matrix of evaluation points.

obj

kdevinecop object.

stable

logical; option for stabilizing the estimator: the estimated pair copula density is cut off at \(50\).

n

integer; number of observations.

U

(optional) \(n x d\) matrix of independent uniform random variables.

quasi

logical; the default (FALSE) returns pseudo-random numbers, use TRUE for quasi-random numbers (generalized Halton, see ghalton).

Value

A numeric vector of the density/cdf or a \(n x 2\) matrix of simulated data.

References

Nagler, T., Czado, C. (2016)
Evading the curse of dimensionality in nonparametric density estimation.
Journal of Multivariate Analysis 151, 69-89 (doi:10.1016/j.jmva.2016.07.003)

Dissmann, J., Brechmann, E. C., Czado, C., and Kurowicka, D. (2013).
Selecting and estimating regular vine copulae and application to financial returns.
Computational Statistics & Data Analysis, 59(0):52--69.

See also

Author

Thomas Nagler

Examples

data(wdbc, package = "kdecopula")                    # load data
u <- VineCopula::pobs(wdbc[, 5:7], ties = "average") # rank-transform

fit <- kdevinecop(u)                # estimate density
dkdevinecop(c(0.1, 0.1, 0.1), fit)  # evaluate density estimate
#> [1] 3.199711