Kernel Smoothing for Bivariate Copula Densities

This package implements a vine copula based kernel density estimator. The estimator does not suffer from the curse of dimensionality and is therefore well suited for high-dimensional applications (see, Nagler and Czado, 2016).

Details

The multivariate kernel density estimators is implemented by the kdevine function. It combines a kernel density estimator for the margins (kde1d) and a kernel estimator of the vine copula density (kdevinecop). The package is built on top of the copula density estimators in the kdecopula::kdecopula-package and let's you choose from all its implemented methods. Optionally, the vine copula can be estimated parameterically (only the margins are nonparametric).

References

Nagler, T., Czado, C. (2016)
Evading the curse of dimensionality in nonparametric density estimation with simplified vine copulas.
Journal of Multivariate Analysis 151, 69-89 (doi:10.1016/j.jmva.2016.07.003)

Nagler, T., Schellhase, C. and Czado, C. (2017)
Nonparametric estimation of simplified vine copula models: comparison of methods arXiv:1701.00845

Nagler, T. (2017)
A generic approach to nonparametric function estimation with mixed data.
arXiv:1704.07457